Move every file into its correct folder in v2
