+++ client/src/components/v2/boost/BoostConfigList.tsx
@@ -4,7 +4,7 @@ import { Button } from 'react-bootstrap';
 import { BoostConfig } from 'types/boost';
 import styles from './BoostConfigList.module.css';
 import FontAwesomeIcon from 'components/v2/FontAwesomeIcon';
-import WidgetListGroupItem from './WidgetListGroupItem';
+import WidgetListGroupItem from '../WidgetListGroupItem';
 
 export interface BoostConfigListProps {
   config: BoostConfig;
