+++ client/src/components/v2/camera_settings/CameraStatus.js
