+++ client/src/components/v2/dashboard/V2TextModeChart.js
