+++ client/src/components/v2/power_map/LabelledControl.js
