+++ client/src/views/v2/DownloadFilesView.js
@@ -1,8 +1,8 @@
 import React, { useState, useCallback } from 'react';
 import { Button } from 'react-bootstrap';
-import ContentPage from 'components/v2/ContentPage';
-import DeleteModal from 'components/v2/DeleteModal';
-import LogFileList from 'components/v2/LogFileList';
+import ContentPage from 'components/common/ContentPage';
+import DeleteModal from 'components/v2/download_files/DeleteModal';
+import LogFileList from 'components/v2/download_files/LogFileList';
 import { useFiles, useLatestFile } from 'api/v2/files';
 
 /**
